requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==24.1.0
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.2.2
click==8.1.7
configparser==7.0.0
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mariadb==1.1.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysql-connector-python==9.0.0
orjson==3.10.3
packaging==24.1
passlib==1.7.4
pydantic==2.7.2
pydantic-extra-types==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.3
Pygments==2.18.0
PyMySQL==1.1.1
pyproject-toml==0.0.10
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
referencing==0.35.1
rich==13.7.1
rpds-py==0.19.1
setuptools==72.1.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
toml==0.10.2
typer==0.12.3
typing_extensions==4.12.0
ujson==5.10.0
uvicorn==0.30.0
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
boto3
cryptography
psycopg2
Geometry
geoalchemy2
pandas
matplotlib
seaborn
numpy
openai==0.28
aiofiles
python-dotenv